Tokens: fix app header not applying theme in dark mode
